data: Move file-review page parsing
The file review renderer builds paginated review models and commands use
that module for page-state helpers during show workflows.

Page parsing and persisted page-spec formatting also live in the renderer,
so command code depends on rendering code for a small selection concern.
That makes the output package boundary harder to read when following
file-review page flow.

This commit addresses that by moving page selection parsing and page-spec
normalization into data.file_review.pages while updating show commands
and the renderer to import that helper module.

The following tests commits will cover the page helper directly and lock
the renderer boundary.
